// src/app/models/provider.ts
export interface Provider {
  id: string;
  providerNameId: string;
  providerFullName: string;
  providerPhone: string;
  providerEmail: string;
  providerAdd1: string;
  providerAdd2: string;
  providerCity: string;
  providerState: string;
  providerZip: string;
  providerType: string;
  providerIsActive: boolean;
}